+++ app/src/components/Footer/Footer.jsx
@@ -39,9 +39,9 @@ function Footer({ ...props }) {
         <p className={classes.right}>
           <span>
             &copy; {1900 + new Date().getYear()}{" "}
-            <a href="https://www.creative-tim.com" className={classes.a}>
-              Creative Tim
-            </a>, made with love for a better web
+            <a href="javascript:void(0)" className={classes.a}>
+              Meal Plan
+            </a>, made with love for life
           </span>
         </p>
       </div>