+++ app/src/views/Notifications/Notifications.jsx
@@ -83,12 +83,12 @@ class Notifications extends React.Component {
           <h4 className={classes.cardTitleWhite}>Notifications</h4>
           <p className={classes.cardCategoryWhite}>
             Handcrafted by our friends from{" "}
-            <a target="_blank" href="https://material-ui-next.com/">
-              Material UI
+            <a target="_blank" href="javascript:void(0)">
+              Meal UI
             </a>{" "}
             and styled by{" "}
-            <a target="_blank" href="https://www.creative-tim.com/">
-              Creative Tim
+            <a target="_blank" href="javascript:void(0)">
+              Meal plan
             </a>. Please checkout the{" "}
             <a href="#pablo" target="_blank">
               full documentation